rename packages add page objects for login and dashboard add dependencies related to selenium into pom.xml
