+++ src/main/java/com/fundacionjala/pivotal/framework/selenium/Chrome.java
@@ -1,4 +1,4 @@
-package com.fundacionjala.pivotal.driver;
+package com.fundacionjala.pivotal.framework.selenium;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
@@ -8,7 +8,7 @@ public class Chrome implements IDriver {
     private static String driverPath = "/lib/";
 
     public WebDriver initDriver() {
-        System.setProperty("webdriver.chrome.driver", driverPath + "chromedriver.exe");
+        System.setProperty("webdriver.chrome.selenium", driverPath + "chromedriver.exe");
         return new ChromeDriver();
     }
 }
