+++ src/main/java/com/fundacionjala/pivotal/framework/selenium/RemoteBrowser.java
@@ -1,15 +1,14 @@
-package com.fundacionjala.pivotal.driver;
+package com.fundacionjala.pivotal.framework.selenium;
 
 import java.net.MalformedURLException;
 import java.net.URL;
 
-import com.fundacionjala.pivotal.PropertiesInfo;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.remote.CapabilityType;
 import org.openqa.selenium.remote.DesiredCapabilities;
 import org.openqa.selenium.remote.RemoteWebDriver;
 
-import static com.fundacionjala.pivotal.PropertiesInfo.getInstance;
+import static com.fundacionjala.pivotal.framework.util.PropertiesInfo.getInstance;
 
 public class RemoteBrowser implements IDriver {
