+++ src/main/java/com/fundacionjala/pivotal/pages/BasePage.java
@@ -1,8 +1,7 @@
-package com.fundacionjala.pivotal;
+package com.fundacionjala.pivotal.pages;
 
 
 
-import com.fundacionjala.pivotal.driver.DriverManager;
 import org.openqa.selenium.StaleElementReferenceException;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
@@ -10,7 +9,7 @@ import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.PageFactory;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-import static com.fundacionjala.pivotal.driver.DriverManager.*;
+import static com.fundacionjala.pivotal.framework.selenium.DriverManager.*;
 
 public abstract class BasePage {
