+++ src/main/java/com/fundacionjala/pivotal/pages/CreateProject.java
@@ -1,0 +1,7 @@
+package com.fundacionjala.pivotal.pages;
+
+/**
+ * Created by mijhailvillarroel on 7/6/2016.
+ */
+public class CreateProject {
+}
