+++ src/main/java/com/fundacionjala/pivotal/pages/Dashboard.java
@@ -1,14 +1,13 @@
-package com.fundacionjala.pivotal;
+package com.fundacionjala.pivotal.pages;
 
 import java.util.concurrent.TimeUnit;
 
-import com.fundacionjala.pivotal.driver.DriverManager;
 import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
-import static com.fundacionjala.pivotal.driver.DriverManager.IMPLICIT_FAIL_WAIT_TIME;
-import static com.fundacionjala.pivotal.driver.DriverManager.IMPLICIT_WAIT_TIME;
+import static com.fundacionjala.pivotal.framework.selenium.DriverManager.IMPLICIT_FAIL_WAIT_TIME;
+import static com.fundacionjala.pivotal.framework.selenium.DriverManager.IMPLICIT_WAIT_TIME;
 
 /**
  *
